requirements/base.txt
Django==3.2.1
djangorestframework==3.12.4
python-decouple==3.4
